+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 
 - Added support for `DataFrame.create_or_replace_view` and `Series.create_or_replace_view`.
 - Added support for `DataFrame.create_or_replace_dynamic_table` and `Series.create_or_replace_dynamic_table`.
+- Added support for `DataFrame.to_view` and `Series.to_view`.
 - Added support for `DataFrame.to_dynamic_table` and `Series.to_dynamic_table`.
 
 #### Improvements